move common_options from core.py to options.py
